requirements.txt
PyPDF2
PyMuPDF
Pillow
easyocr
python-dotenv
Flask
docx2pdf
fpdf
langchain
langchain-google-genai
scikit-learn
numpy
flask-cors
